Crate smbclient_sys

Structs

_SMBCCTX

_smbc_callbacks

_smbc_options

print_job_info

smbc_dirent

Enums

SMBC_internal_data

_SMBCFILE

_SMBCSRV

smbc_server_cache

Constants

SMBC_ENCRYPTLEVEL_NONE

SMBC_ENCRYPTLEVEL_REQUEST

SMBC_ENCRYPTLEVEL_REQUIRE

SMBC_SHAREMODE_DENY_ALL

SMBC_SHAREMODE_DENY_DOS

SMBC_SHAREMODE_DENY_FCB

SMBC_SHAREMODE_DENY_NONE

SMBC_SHAREMODE_DENY_READ

SMBC_SHAREMODE_DENY_WRITE

SMBC_VFS_FEATURE_CASE_INSENSITIVE

SMBC_VFS_FEATURE_DFS

SMBC_VFS_FEATURE_NO_UNIXCIFS

SMBC_VFS_FEATURE_RDONLY

Functions

smbc_chmod

smbc_close

smbc_closedir

smbc_creat

smbc_fgetxattr

smbc_flistxattr

smbc_free_context

smbc_fremovexattr

smbc_fsetxattr

smbc_fstat

smbc_fstatvfs

smbc_ftruncate

smbc_getDebug

smbc_getFunctionAddCachedServer

smbc_getFunctionAuthData

smbc_getFunctionAuthDataWithContext

smbc_getFunctionCheckServer

smbc_getFunctionChmod

smbc_getFunctionClose

smbc_getFunctionClosedir

smbc_getFunctionCreat

smbc_getFunctionFstat

smbc_getFunctionFstatVFS

smbc_getFunctionFstatdir

smbc_getFunctionFtruncate

smbc_getFunctionGetCachedServer

smbc_getFunctionGetdents

smbc_getFunctionGetxattr

smbc_getFunctionListPrintJobs

smbc_getFunctionListxattr

smbc_getFunctionLseek

smbc_getFunctionLseekdir

smbc_getFunctionMkdir

smbc_getFunctionOpen

smbc_getFunctionOpenPrintJob

smbc_getFunctionOpendir

smbc_getFunctionPrintFile

smbc_getFunctionPurgeCachedServers

smbc_getFunctionRead

smbc_getFunctionReaddir

smbc_getFunctionRemoveCachedServer

smbc_getFunctionRemoveUnusedServer

smbc_getFunctionRemovexattr

smbc_getFunctionRename

smbc_getFunctionRmdir

smbc_getFunctionSetxattr

smbc_getFunctionStat

smbc_getFunctionStatVFS

smbc_getFunctionTelldir

smbc_getFunctionUnlink

smbc_getFunctionUnlinkPrintJob

smbc_getFunctionUtimes

smbc_getFunctionWrite

smbc_getNetbiosName

smbc_getOptionBrowseMaxLmbCount

smbc_getOptionCaseSensitive

smbc_getOptionDebugToStderr

smbc_getOptionFallbackAfterKerberos

smbc_getOptionFullTimeNames

smbc_getOptionNoAutoAnonymousLogin

smbc_getOptionOneSharePerServer

smbc_getOptionOpenShareMode

smbc_getOptionSmbEncryptionLevel

smbc_getOptionUrlEncodeReaddirEntries

smbc_getOptionUseCCache

smbc_getOptionUseKerberos

smbc_getOptionUserData

smbc_getServerCacheData

smbc_getTimeout

smbc_getUser

smbc_getWorkgroup

smbc_getdents

smbc_getxattr

smbc_init

smbc_init_context

smbc_lgetxattr

smbc_list_print_jobs

smbc_listxattr

smbc_llistxattr

smbc_lremovexattr

smbc_lseek

smbc_lseekdir

smbc_lsetxattr

smbc_mkdir

smbc_new_context

smbc_open

smbc_open_print_job

smbc_opendir

smbc_option_get

smbc_option_set

smbc_print_file

smbc_read

smbc_readdir

smbc_remove_unused_server

smbc_removexattr

smbc_rename

smbc_rmdir

smbc_setDebug

smbc_setFunctionAddCachedServer

smbc_setFunctionAuthData

smbc_setFunctionAuthDataWithContext

smbc_setFunctionCheckServer

smbc_setFunctionChmod

smbc_setFunctionClose

smbc_setFunctionClosedir

smbc_setFunctionCreat

smbc_setFunctionFstat

smbc_setFunctionFstatVFS

smbc_setFunctionFstatdir

smbc_setFunctionFtruncate

smbc_setFunctionGetCachedServer

smbc_setFunctionGetdents

smbc_setFunctionGetxattr

smbc_setFunctionListPrintJobs

smbc_setFunctionListxattr

smbc_setFunctionLseek

smbc_setFunctionLseekdir

smbc_setFunctionMkdir

smbc_setFunctionOpen

smbc_setFunctionOpenPrintJob

smbc_setFunctionOpendir

smbc_setFunctionPrintFile

smbc_setFunctionPurgeCachedServers

smbc_setFunctionRead

smbc_setFunctionReaddir

smbc_setFunctionRemoveCachedServer

smbc_setFunctionRemoveUnusedServer

smbc_setFunctionRemovexattr

smbc_setFunctionRename

smbc_setFunctionRmdir

smbc_setFunctionSetxattr

smbc_setFunctionStat

smbc_setFunctionStatVFS

smbc_setFunctionTelldir

smbc_setFunctionUnlink

smbc_setFunctionUnlinkPrintJob

smbc_setFunctionUtimes

smbc_setFunctionWrite

smbc_setNetbiosName

smbc_setOptionBrowseMaxLmbCount

smbc_setOptionCaseSensitive

smbc_setOptionDebugToStderr

smbc_setOptionFallbackAfterKerberos

smbc_setOptionFullTimeNames

smbc_setOptionNoAutoAnonymousLogin

smbc_setOptionOneSharePerServer

smbc_setOptionOpenShareMode

smbc_setOptionSmbEncryptionLevel

smbc_setOptionUrlEncodeReaddirEntries

smbc_setOptionUseCCache

smbc_setOptionUseKerberos

smbc_setOptionUserData

smbc_setServerCacheData

smbc_setTimeout

smbc_setUser

smbc_setWorkgroup

smbc_set_context

smbc_set_credentials

smbc_set_credentials_with_fallback

smbc_setxattr

smbc_stat

smbc_statvfs

smbc_telldir

smbc_thread_impl

smbc_thread_posix

smbc_unlink

smbc_unlink_print_job

smbc_urldecode

smbc_urlencode

smbc_utimes

smbc_version

smbc_write

Type Definitions

SMBCCTX

SMBCFILE

SMBCSRV

smbc_add_cached_srv_fn

smbc_bool

smbc_check_server_fn

smbc_chmod_fn

smbc_close_fn

smbc_closedir_fn

smbc_creat_fn

smbc_fstat_fn

smbc_fstatdir_fn

smbc_fstatvfs_fn

smbc_ftruncate_fn

smbc_get_auth_data_fn

smbc_get_auth_data_with_context_fn

smbc_get_cached_srv_fn

smbc_getdents_fn

smbc_getxattr_fn

smbc_list_print_job_fn

smbc_list_print_jobs_fn

smbc_listxattr_fn

smbc_lseek_fn

smbc_lseekdir_fn

smbc_mkdir_fn

smbc_open_fn

smbc_open_print_job_fn

smbc_opendir_fn

smbc_print_file_fn

smbc_purge_cached_fn

smbc_read_fn

smbc_readdir_fn

smbc_remove_cached_srv_fn

smbc_remove_unused_server_fn

smbc_removexattr_fn

smbc_rename_fn

smbc_rmdir_fn

smbc_setxattr_fn

smbc_share_mode

smbc_smb_encrypt_level
Values for option SMB Encryption Level, as set and retrieved with smbc_setOptionSmbEncryptionLevel() and smbc_getOptionSmbEncryptionLevel()

smbc_stat_fn

smbc_statvfs_fn

smbc_telldir_fn

smbc_unlink_fn

smbc_unlink_print_job_fn

smbc_utimes_fn

smbc_vfs_feature
Capabilities set in the f_flag field of struct statvfs, from smbc_statvfs(). These may be OR-ed together to reflect a full set of available capabilities.

smbc_write_fn