Struct smbclient_sys::print_job_info

pub struct print_job_info {
    pub id: c_ushort,
    pub priority: c_ushort,
    pub size: size_t,
    pub user: [c_char; 128usize],
    pub name: [c_char; 128usize],
    pub t: time_t,
}

Fields

id

numeric ID of the print job

priority

represents print job priority (lower numbers mean higher priority)

size

Size of the print job

user

Name of the user that owns the print job

name

Name of the print job. This will have no name if an anonymous print * file was opened. Ie smb://server/printer

t

Time the print job was spooled