Struct smbclient_sys::smbc_dirent

pub struct smbc_dirent {
    pub smbc_type: c_uint,
    pub dirlen: c_uint,
    pub commentlen: c_uint,
    pub comment: *mut c_char,
    pub namelen: c_uint,
    pub name: [c_char; 1usize],
}

Fields

smbc_type

Type of entity. SMBC_WORKGROUP=1, SMBC_SERVER=2, SMBC_FILE_SHARE=3, SMBC_PRINTER_SHARE=4, SMBC_COMMS_SHARE=5, SMBC_IPC_SHARE=6, SMBC_DIR=7, SMBC_FILE=8, SMBC_LINK=9,

dirlen

Length of this smbc_dirent in bytes

commentlen

The length of the comment string in bytes (does not include * null terminator)

comment

Points to the null terminated comment string

namelen

The length of the name string in bytes (does not include * null terminator)

name

Points to the null terminated name string